+++ aiida_vasp/commands/tests/test_mock_vasp.py
@@ -5,4 +5,4 @@ import subprocess
 
 def test_mock_vasp():
     """A simple test to verify that mock-vasp can run."""
-    subprocess.check_output(["mock-vasp", "--help"])
+    subprocess.check_output(['mock-vasp', '--help'])