+++ examples/run_bands.py
@@ -1,72 +1,130 @@
-import click
-import os
+"""
+An example call script that performs a single static VASP calculation.
+
+Performs a self consistent electron convergence run using the standard silicon structure.
+"""
+# pylint: disable=too-many-arguments
+import numpy as np
 from aiida.common.extendeddicts import AttributeDict
-from aiida.cmdline.utils.decorators import with_dbenv
+from aiida.orm import Code, Bool, Str
+from aiida.plugins import DataFactory, WorkflowFactory
+from aiida.engine import submit
+from aiida import load_profile
+load_profile()
 
-from aiida_vasp.utils.aiida_utils import get_data_node
-from auxiliary import example_param_set, set_structure_si, set_kpoints, set_params_simple, set_params_simple_no_encut
 
-os.system('verdi daemon restart')
+def get_structure():
+    """
+    Set up Si primitive cell
+
+    Si
+       5.431
+         0.0000000000000000    0.5000000000000000    0.5000000000000000
+         0.5000000000000000    0.0000000000000000    0.5000000000000000
+         0.5000000000000000    0.5000000000000000    0.0000000000000000
+    Si
+       2
+    Direct
+      0.8750000000000000  0.8750000000000000  0.8750000000000000
+      0.1250000000000000  0.1250000000000000  0.1250000000000000
+
+    """
+
+    structure_data = DataFactory('structure')
+    alat = 5.431
+    lattice = np.array([[.5, 0, .5], [.5, .5, 0], [0, .5, .5]]) * alat
+    structure = structure_data(cell=lattice)
+    for pos_direct in ([0.875, 0.875, 0.875], [0.125, 0.125, 0.125]):
+        pos_cartesian = np.dot(pos_direct, lattice)
+        structure.append_atom(position=pos_cartesian, symbols='Si')
+    return structure
 
 
-@click.command()
-@example_param_set
-@with_dbenv()
-def main(potential_family, queue, code, computer):
-    from aiida.orm import Code
-    from aiida.plugins import WorkflowFactory
-    from aiida.engine import submit, run
-    from aiida_vasp.utils.aiida_utils import get_data_class
+def main(code_string, incar, kmesh, structure, potential_family, potential_mapping, options):
+    """Main method to setup the calculation."""
 
-    # set the code to be used, currently tied to a computer
-    code = Code.get_from_string('{}@{}'.format(code, computer))
+    # First, we need to fetch the AiiDA datatypes which will
+    # house the inputs to our calculation
+    dict_data = DataFactory('dict')
+    kpoints_data = DataFactory('array.kpoints')
 
-    # set the workchain you would like to call
+    # Then, we set the workchain you would like to call
     workchain = WorkflowFactory('vasp.master')
 
-    # organize options (needs a bit of special care)
-    options = AttributeDict()
-    if computer == 'unity':
-        options.account = ''
-        options.qos = ''
-        options.resources = {'num_machines': 1, 'num_mpiprocs_per_machine': 20}
-    elif computer == 'fram':
-        options.account = 'nn9544k'
-        options.qos = 'devel'
-        options.resources = {'num_machines': 1, 'num_mpiprocs_per_machine': 16}
-    options.queue_name = ''
-    options.max_wallclock_seconds = 3600
-
-    # organize settings
+    # And finally, we declare the options, settings and input containers
     settings = AttributeDict()
-    settings.parser_settings = AttributeDict({'output_params': ['total_energies', 'maximum_force']})
-
-    # set inputs for the following WorkChain execution
     inputs = AttributeDict()
-    # set code
-    inputs.code = code
-    # set structure
-    inputs.structure = set_structure_si()
-    # set k-points grid density
-    inputs.kpoints = set_kpoints(inputs.structure)
-    # set parameters
-    inputs.parameters = set_params_simple()
-    # set potentials and their mapping
-    inputs.potential_family = get_data_class('str')(potential_family)
-    inputs.potential_mapping = get_data_node('dict', dict={'Si': 'Si'})
-    # set options
-    inputs.options = get_data_node('dict', dict=options)
-    # set settings
-    inputs.settings = get_data_node('dict', dict=settings)
-    # set workchain related inputs
-    inputs.verbose = get_data_node('bool', True)
-    # do we want to relax?
-    inputs.relax = get_data_node('bool', False)
-    # do we want to extract the band structure?
-    inputs.extract_bands = get_data_node('bool', True)
-    # submit the requested workchain with the supplied inputs
+
+    # Organize settings
+    settings.parser_settings = {'output_params': ['total_energies', 'maximum_force']}
+
+    # Set inputs for the following WorkChain execution
+    # Set code
+    inputs.code = Code.get_from_string(code_string)
+    # Set structure
+    inputs.structure = structure
+    # Set k-points grid density
+    kpoints = kpoints_data()
+    kpoints.set_kpoints_mesh(kmesh)
+    inputs.kpoints = kpoints
+    # Set parameters
+    inputs.parameters = dict_data(dict=incar)
+    # Set potentials and their mapping
+    inputs.potential_family = Str(potential_family)
+    inputs.potential_mapping = dict_data(dict=potential_mapping)
+    # Set options
+    inputs.options = dict_data(dict=options)
+    # Set settings
+    inputs.settings = dict_data(dict=settings)
+    # Set workchain related inputs, in this case, give more explicit output to repor
+    inputs.verbose = Bool(True)
+
+    # Master, convergence and relaxation related parameters that is passed to the master,
+    # convergence and relaxation workchain, respectively
+    # Turn of relaxation
+    inputs.relax = Bool(False)
+    # Extract electronic band structure
+    inputs.extract_bands = Bool(True)
+    # Submit the requested workchain with the supplied inputs
     submit(workchain, **inputs)
 
 
 if __name__ == '__main__':
-    main()
+    # Code_string is chosen among the list given by 'verdi code list'
+    CODE_STRING = 'vasp@mycluster'
+
+    # POSCAR equivalent
+    # Set the silicon structure
+    STRUCTURE = get_structure()
+
+    # INCAR equivalent
+    # Set input parameters
+    INCAR = {'prec': 'NORMAL', 'encut': 200, 'ediff': 1E-4, 'ialgo': 38, 'ismear': -5, 'sigma': 0.1}
+
+    # KPOINTS equivalent
+    # Set kpoint mesh
+    KMESH = [9, 9, 9]
+
+    # POTCAR equivalent
+    # Potential_family is chosen among the list given by
+    # 'verdi data vasp-potcar listfamilies'
+    POTENTIAL_FAMILY = 'pbe'
+    # The potential mapping selects which potential to use, here we use the standard
+    # for silicon, this could for instance be {'Si': 'Si_GW'} to use the GW ready
+    # potential instead
+    POTENTIAL_MAPPING = {'Si': 'Si'}
+
+    # Jobfile equivalent
+    # In options, we typically set scheduler options.
+    # See https://aiida.readthedocs.io/projects/aiida-core/en/latest/scheduler/index.html
+    # AttributeDict is just a special dictionary with the extra benefit that
+    # you can set and get the key contents with mydict.mykey, instead of mydict['mykey']
+    OPTIONS = AttributeDict()
+    OPTIONS.account = ''
+    OPTIONS.qos = ''
+    OPTIONS.resources = {'num_machines': 1, 'num_mpiprocs_per_machine': 16}
+    OPTIONS.queue_name = ''
+    OPTIONS.max_wallclock_seconds = 3600
+    OPTIONS.max_memory_kb = 1024000
+
+    main(CODE_STRING, INCAR, KMESH, STRUCTURE, POTENTIAL_FAMILY, POTENTIAL_MAPPING, OPTIONS)